changed the pagination style on assets tables
