+++ resources/js/components/shared/index.js
@@ -54,6 +54,7 @@ import DataLoading from '../../components/common/DataLoading';
 import AvatarImage from '../../components/AvatarImage';
 import FilterTableBodyMixin from "./FilterTableBodyMixin";
 import PaginationTable from "./PaginationTable.vue";
+import { ellipsisSortClick } from "./UtilsTable";
 
 export {
   AddToProjectModal,
@@ -112,4 +113,5 @@ export {
   AvatarImage,
   FilterTableBodyMixin,
   PaginationTable,
+  ellipsisSortClick,
 };
