+++ resources/js/processes-catalogue/components/RequestTab.vue
@@ -132,7 +132,8 @@ export default {
       this.page = page;
       this.queryBuilder();
     },
-    changePerPage(value) {
+    changePerPage(value, page) {
+      this.page = page;
       this.perPage = value;
       this.queryBuilder();
     },
